requirements.txt
Flask==3.1.0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
obs-websocket-py==1.0
sseclient==0.0.27
pillow==11.2.1
requests==2.32.3
websocket-client==1.8.0